save rendering pipeline stages
The rendering pipeline stage is the largest I had to set up so far. It
has the most steps and is as large as the initial stages for the whole
triangle rendering in vulkan. Figuring out what I need to do
conceptually on abstract level and figuring out intricacies helps me
manage the complexity.